Обновить стили для кнопок. Добавить страницу help.html. Добавить новые страницы в sitemap.txt
