+++ main_app/static/main_app/css/help.css
@@ -1,0 +1,76 @@
+span {
+    font-size: 12px;
+}
+
+.content {
+    text-align: center;
+}
+
+.inline {
+    display: inline-block;
+    width: 680px;
+}
+
+.txt,
+.txt-area {
+    display: inline-block;
+    color: #676767;
+    margin-bottom: 10px;
+    border: 1px solid #ccc;
+    padding: 5px 10px;
+    font-size: 15px!important;
+    line-height: 1.4em;
+}
+
+.txt {
+   width: 200px;
+}
+
+.txt-area {
+   width: 653px;
+   min-height: 150px;
+   max-height: 400px;
+   resize: vertical;
+}
+
+.txt:focus,
+.txt-area:focus {
+   border-style: solid;
+   border-color: #bababa;
+   color: #444;
+}
+
+input.error,
+textarea.error {
+    border-color: #973D3D;
+    border-style: solid;
+    background: #F0BEBE;
+    color: #A35959;
+}
+
+input.error:focus,
+textarea.error:focus {
+    border-color: #973D3D;
+    color: #A35959;
+}
+
+.send {
+    color: #fff;
+    cursor: pointer;
+    margin: 20px 0;
+    padding: 10px;
+    font-size: 12px;
+    font-weight: 600;
+    border: solid 1px #f9f9f9;
+    border-radius: 5px;
+    background: #3F51B5;
+    width: 680px;
+}
+
+.send:hover {
+    background: #E27715;
+}
+
+.footer {
+    position: absolute;
+}
