+++ main_app/static/main_app/css/index.css
@@ -8,11 +8,15 @@ p.address {
     margin: 0;
 }
 
-input.text_address {
-    padding: 8px;
+input.text_address,
+input.button_address {
     font-weight: 600;
     font-size: 18px;
     border-radius: 5px;
+}
+
+input.text_address {
+    padding: 8px;
     width: 500px;
 }
 
@@ -21,12 +25,14 @@ input.button_address {
     -webkit-appearance:none;
     padding: 10px;
     border: none;
-    background-color: #3F51B5;
-    color: #fff;
-    font-weight: 600;
-    font-size: 18px;
-    border-radius: 5px;
+    background: #3F51B5;
+    color: #FFF;
     width: 120px;
+}
+
+input.button_address:hover {
+    background: #E27715;
+    cursor: pointer;
 }
 
 .footer {
