+++ main_app/templates/main_app/base.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ru">
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-        <title>{% block title %}{% endblock %}</title>
+        <title>Pets Friendly &#124; {% block title %}{% endblock %}</title>
         <meta name="description" content="{% block description %}{% endblock %}" />
         <link rel="icon" href="https://pets-friendly.ru/favicon.svg" type="image/svg+xml">
         {% load static %}
@@ -35,7 +35,7 @@
                     <div class="col-4">
                         <ul class="footer-navigation-link-list">
                             <li class="footer-navigation-link">
-                                <a href="/">Связаться с нами</a>
+                                <a href="{% url 'help' %}">Связаться с нами</a>
                             </li>
                         </ul>
                     </div>
