+++ main_app/templates/main_app/hotels.html
@@ -1,7 +1,7 @@
 {% extends 'main_app/base.html' %}
 {% load static %}
 <!--TODO: реализовать динмический тайтл при выборе города-->
-{% block title %}Pets Friendly: Отели для питомцев в Москве{% endblock %}
+{% block title %}Отели для питомцев в Москве{% endblock %}
 {% block description %}
 Более 100 отелей в Москве и Московской области. Гарантия лучшей цены! Без комиссий. Официальный сайт.
 {% endblock %}
