+++ main_app/urls.py
@@ -6,5 +6,6 @@ from . import views
 urlpatterns = [
     path('', views.index_page, name='index'),
     path('hotels/', views.hotels_page, name='hotels'),
-    path('hotels/<int:hotel_id>/', views.card_hotel_page, name='card_hotel'),
+    # path('hotels/<int:hotel_id>/', views.card_hotel_page, name='card_hotel'),
+    path('help/', views.help_page, name='help'),
 ]
