+++ sitemap.txt
@@ -1,2 +1,3 @@
 https://pets-friendly.ru/
-https://www.pets-friendly.ru/
+https://pets-friendly.ru/hotels
+https://pets-friendly.ru/help
